requirements.txt
asn1crypto==1.5.1
Flask==3.0.3
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
cryptography==43.0.3
oscrypto==1.3.0
qrcode==8.0
reportlab==4.2.5
python-dotenv==1.0.0
pymongo
confluent-kafka
azure-storage-blob
azure-identity
